requirements.txt
streamlit
openai
pyyaml
requests
python-dotenv
